Separating app.py into control.py
